Show error dialog when device doesn’t have map app
